Move stac reader into it's own file and write tests for _get_reader
