+++ tests/fixtures/C3S-LC-L4-LCCS-Map-300m-P1Y-2020-v2.1.1.json
@@ -1,0 +1,1 @@
+{"id":"C3S-LC-L4-LCCS-Map-300m-P1Y-2020-v2.1.1","bbox":[-180.0,-90.0,180.0,90.0],"type":"Feature","links":[{"rel":"collection","type":"application/json","href":"https://planetarycomputer.microsoft.com/api/stac/v1/collections/esa-cci-lc-netcdf"},{"rel":"parent","type":"application/json","href":"https://planetarycomputer.microsoft.com/api/stac/v1/collections/esa-cci-lc-netcdf"},{"rel":"root","type":"application/json","href":"https://planetarycomputer.microsoft.com/api/stac/v1/"},{"rel":"self","type":"application/geo+json","href":"https://planetarycomputer.microsoft.com/api/stac/v1/collections/esa-cci-lc-netcdf/items/C3S-LC-L4-LCCS-Map-300m-P1Y-2020-v2.1.1"}],"assets":{"netcdf":{"href":"https://landcoverdata.blob.core.windows.net/esa-cci-lc/netcdf/C3S-LC-L4-LCCS-Map-300m-P1Y-2020-v2.1.1.nc","created":"2021-05-18T10:15:40Z","cube:variables":{"crs":{"type":"auxiliary","dimensions":[]},"lat":{"axis":"Y","type":"auxiliary","unit":"degrees_north","dimensions":["lat"],"description":"latitude"},"lon":{"axis":"X","type":"auxiliary","unit":"degrees_east","dimensions":["lon"],"description":"longitude"},"time":{"axis":"T","type":"auxiliary","unit":"days since 1970-01-01 00:00:00","dimensions":["time"],"description":"time"},"lat_bounds":{"type":"auxiliary","dimensions":["lat","bounds"]},"lccs_class":{"type":"data","dimensions":["time","lat","lon"],"description":"Land cover class defined in LCCS"},"lon_bounds":{"type":"auxiliary","dimensions":["lon","bounds"]},"time_bounds":{"type":"auxiliary","dimensions":["time","bounds"]},"change_count":{"type":"data","dimensions":["time","lat","lon"],"description":"number of class changes"},"processed_flag":{"type":"data","dimensions":["time","lat","lon"],"description":"LC map processed area flag"},"observation_count":{"type":"data","dimensions":["time","lat","lon"],"description":"number of valid observations"},"current_pixel_state":{"type":"data","dimensions":["time","lat","lon"],"description":"LC pixel type mask"}},"cube:dimensions":{"lat":{"type":"lat","extent":[-90.0,90.0]},"lon":{"type":"lon","extent":[-180.0,180.0]},"time":{"type":"time","values":[18262.0]},"bounds":{"type":"bounds","values":[0,1]}},"type":"application/netcdf","roles":["data","quality"],"title":"ESA CCI Land Cover NetCDF 4 File"}},"geometry":{"type":"Polygon","coordinates":[[[-180,-90],[180,-90],[180,90],[-180,90],[-180,-90]]]},"collection":"esa-cci-lc-netcdf","properties":{"title":"ESA CCI Land Cover Map for 2020","created":"2023-01-11T18:52:07.171349Z","datetime":null,"proj:epsg":4326,"proj:shape":[129600,64800],"end_datetime":"2020-12-31T23:59:59Z","proj:transform":[-180.0,0.002777777777778,0.0,90.0,0.0,-0.002777777777778],"start_datetime":"2020-01-01T00:00:00Z","esa_cci_lc:version":"2.1.1","processing:lineage":"Produced based on the following data sources: Sentinel-3 OLCI","processing:software":{"lc-sr":"1.0","lc-user-tools(1)":"3.14","lc-user-tools(2)":"4.5","lc-classification":"1.0"}},"stac_extensions":["https://stac-extensions.github.io/projection/v1.0.0/schema.json","https://stac-extensions.github.io/processing/v1.1.0/schema.json","https://stac-extensions.github.io/datacube/v2.1.0/schema.json"],"stac_version":"1.0.0"}
